Fix overshoot issue in F.to_pil_image
When converting a tensor to a PIL image, overshoots are not clamped. This means that a value of 1.001 becomes 0 instead of 255. This issue is described here: https://github.com/pytorch/vision/issues/2950#issuecomment-722070283
Additionally, replace the imprecise truncation by a rounding.
Both issues are particularly annoying for downsampling/upsampling code, where such overshoots are common.